Configure auth: Keep user logged in
Netlify identity is based on two tokens: a long-lived refresh token and
a short-lived JSON web token (JWT).

The refresh token is valid for several days, if not weeks, and is used
to create JWTs. The refresh token must be very carefully guarded, since
it has such a long expiry date and can be used to create JWTs again and
again. We store this token in local storage.

The JWT is valid for one hour, is set as a cookie, and is used to
authenticate the user in any SSR pages/functions.

While the user is active on the app, we run an interval in the
background every 5 minutes, that requests a new JWT when the current one
is close to expiry (less than 15 mins remaining).

When the user returns to the app after a break longer than one hour,
the last JWT will no longer be valid and the user will briefly see the
login page. Once the login page loads, we check for the refresh token in
local storage and then request a new JWT. If that succeeds, the user is
automatically logged in and redirected to the home page.