Review: Bypass location and score fields for now
Admins do not review location and SDG score fields for now. They are 
lower in priority and we do not yet have the functionality for reviewing 
two fields at once (latitude & longitude) or devised a good approach to 
reviewing the 17 SDGs. Bypassed fields are stored in FaunaDB, so we will 
need to do a hash-bust (change the extraction hash for all orgs, so that 
they are queued for re-review) and we can then read the review_bypass 
field to see which fields need to undergo human review even though there 
is a match between reviewed value and extracted value.