Review: Always display all steps, even those not needing review
Sometimes users may want to modify fields that don't need review. For 
example, a user may want to upload an organization logo as part of a 
review process, even when the AI has not extracted a logo. Or the user 
may see the need for change in a certain value, even when the field has 
not been updated by the AI.