On complete review, overwrite existing reviewed_extraction_data
When user performs complete review, we clear the current reviewed
extraction data before setting the new one. Essentially, this serves as
a replace operation. By default, q.Update will merge the properties of
two objects, so if we did not clear it first, we'd end up with old/stale
data in the object.

See https://forums.fauna.com/t/replace-by-path/169/2 for context.